refactor: 패키지 이동 dto -> dto/workbook // workResponse 필드 수정
